Crate ibc_middleware_packet_forward

Rust implementation of the IBC packet forwarding middleware.

Structs

Duration

Duration type whose serialization routines are compatible with Strangelove’s PFM JSON forward messages.

ForwardMetadata

Metadata included in ICS-20 packet memos, related with the packet forwarding middleware.

InFlightPacket

Packet that is currently being transmitted to a destination chain over multiple hops.

InFlightPacketKey

InFlightPacket data to facilitate its storage in a key:value store.

PacketForwardMiddleware

Packet forward middleware entrypoint, which intercepts compatible ICS-20 packets and forwards them to other chains.

PacketMetadata

Metadata included in ICS-20 packet memos.

Traits

PfmContext

Context data required by the PacketForwardMiddleware.

Type Aliases

RetryInFlightPacket

Type alias for an InFlightPacket, to enhance code readability.