ibc_middleware_packet_forward
Type Alias RetryInFlightPacket

pub type RetryInFlightPacket = InFlightPacket;

Type alias for an InFlightPacket, to enhance code readability.

Aliased Type

struct RetryInFlightPacket {
    pub original_sender_address: Signer,
    pub refund_port_id: PortId,
    pub refund_channel_id: ChannelId,
    pub packet_src_port_id: PortId,
    pub packet_src_channel_id: ChannelId,
    pub packet_timeout_timestamp: TimeoutTimestamp,
    pub packet_timeout_height: TimeoutHeight,
    pub packet_data: Vec<u8>,
    pub refund_sequence: Sequence,
    pub retries_remaining: Option<NonZero<u8>>,
    pub timeout: Duration,
}

Fields
original_sender_address: Signer

Sender of the packet on the source chain.

refund_port_id: PortId

Port where the packet was received in the current chain.

refund_channel_id: ChannelId

Channel where the packet was received in the current chain.

packet_src_port_id: PortId

Port on the sending chain.

packet_src_channel_id: ChannelId

Channel on the sending chain.

packet_timeout_timestamp: TimeoutTimestamp

Timeout timestamp of the original packet.

packet_timeout_height: TimeoutHeight

Timeout height of the original packet.

packet_data: Vec<u8>

Data of the source packet.

refund_sequence: Sequence

Sequence number of the source packet.

retries_remaining: Option<NonZero<u8>>

Number of retries remaining before the packet is refunded.

timeout: Duration

Timeout duration, relative to some instant (usually a block timestamp).